Move AddressError to dedicated error.rs module
